Add files via upload
I'm bored with just basic counting. How about we do some complex things this time? But I don't want to be associated with any number that's not divisible by 3.
Think you can handle this?

Instructions:
Print out all numbers from 1 to 100 that is divisible by 3, each in separate lines, using a for() loop.