Add files via upload
You have been cordially invited to partake in Operation: Operation. Your mission, should you choose to accept it, is to take the two numbers and the operator given and then perform the operation successfully.

Instructions:
Input one number (integer or decimal), an operator (+, -, *, /), and another number (integer or decimal).
Print the result of the operation between the two numbers, up to 2 decimal places.