Create Area_of_Pentagon.cs
Area of a Pentagon
by CodeChum Admin

Write a program that prompts the user to enter the length of the side of a regular pentagon. The program should use the predefined math functions to calculate the area of the pentagon using the entered side length. Finally, the program should print the calculated area of the pentagon in two decimal places.

 

Area of pentagon formula:



Where a is the length of one side

Sample Output 1
Enter side: 5
Area of Pentagon: 43.01

Sample Output 2
Enter side: 10
Area of Pentagon: 172.05

Sample Output 3
Enter side: 3
Area of Pentagon: 15.48